@@ -2,7 +2,7 @@
 authors = [
 	"9names",
 ]
-edition = "2018"
+edition = "2021"
 readme = "README.md"
 name = "msp432-launchpad-for-ethernet"
 version = "0.1.0"
