@@ -2,7 +2,7 @@
 authors = [
 	"Jonathan 'theJPster' Pallant <github@thejpster.org.uk>",
 ]
-edition = "2018"
+edition = "2021"
 readme = "README.md"
 name = "tiva-c-launchpad"
 version = "0.1.0"
