@@ -12,7 +12,7 @@ keywords = ["arm", "cortex-m", "tm4c", "hal"]
 categories = ["embedded", "hardware-support", "no-std"]
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/rust-embedded-community/tm4c-hal/tm4c129x-hal"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 cortex-m = "0.7"
